Backend/requirements.txt
Django
djangorestframework
django-environ
PyMySQL
Pillow
djangorestframework-simplejwt